Cast to double

+++ lib/shared_components/img_avatar.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 SizedBox restaurantAvatar(dynamic width, dynamic height) {
   return SizedBox(
-      height: height,
-      width: width,
+      height: height.toDouble(),
+      width: width.toDouble(),
       child: ClipRRect(
           borderRadius: BorderRadius.circular(15.0),
           child: FittedBox(
